requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
autopep8==2.0.1
backcall==0.1.0
bleach==3.1.1
blinker==1.4
bottle==0.12.23
cachetools==5.2.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
cloud-init==22.4.2
colorama==0.4.3
command-not-found==0.3
conan==1.55.0
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cycler==0.11.0
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
fasteners==0.18
flatbuffers==23.1.4
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==6.0.0
incremental==16.10.1
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.15.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-console==6.0.0
jupyter-core==4.6.3
keras==2.11.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.3
mistune==0.8.4
more-itertools==4.2.0
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
node-semver==0.6.1
notebook==6.0.3
numpy==1.24.1
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
pandocfilters==1.4.2
parso==0.5.2
patch-ng==1.17.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.4.0
pluginbase==1.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.19.6
psycopg2-binary==2.9.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.10.0
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
PySimpleGUI==4.60.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.7.1
PyYAML==5.3.1
pyzmq==18.1.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0.post1
sos==4.4
ssh-import-id==5.10
systemd-python==234
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
terminado==0.8.2
testpath==0.4.4
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
tornado==5.1.1
tqdm==4.64.1
traitlets==4.3.3
Twisted==18.9.0
typing-extensions==4.4.0
ubuntu-advantage-tools==27.12
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.13
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==2.0.0
wrapt==1.14.1
zipp==1.0.0
zope.interface==4.7.1
